refactor: move inside `python` folder
