+++ src/script/src/lib.rs
@@ -14,11 +14,8 @@
 
 pub mod engine;
 pub mod error;
-mod ffi_types;
 #[cfg(feature = "python")]
 pub mod manager;
-mod pyo3;
 #[cfg(feature = "python")]
 pub mod python;
-mod rspython;
 mod table;
