+++ src/script/src/python.rs
@@ -26,3 +26,7 @@ mod vector;
 
 pub use self::engine::{PyEngine, PyScript};
 pub use self::vector::PyVector;
+
+mod ffi_types;
+mod pyo3;
+mod rspython;
