+++ src/script/src/python/ffi_types/copr/compile.rs
@@ -20,7 +20,7 @@ use rustpython_parser::ast::{Located, Location};
 use rustpython_parser::{ast, parser};
 use snafu::ResultExt;
 
-use crate::ffi_types::copr::parse::{ret_parse_error, DecoratorArgs};
+use crate::python::ffi_types::copr::parse::{ret_parse_error, DecoratorArgs};
 use crate::python::error::{PyCompileSnafu, PyParseSnafu, Result};
 use crate::rspy_fail_parse_error;
